gestion des doublons

+++ script/branch/gestion_dossier.py
@@ -4,7 +4,7 @@ import pandas as pd
 import os
 import shutil
 
-# from ..core.livre import Livre
+from ..leaf.futile import demander_choix_binaire
 
 class GestionDossier():
     def __init__(self, livre):
@@ -20,22 +20,21 @@ class GestionDossier():
 
     def delete_folder(self):
         '''Supprime les dossiers temporaire.'''
-
-        choix = input('Voulez-vous supprimer tous les dossiers commençant par {self.livre.PREFIXE_DOSSIER_TEMPORAIRE} ? (O/N) ').upper()
-        if choix == 'O':
-            # Prend tous les dossiers commençant par vars
-            dossiers = glob.glob(f'{self.livre.PREFIXE_DOSSIER_TEMPORAIRE}*')
-        elif choix == 'N':
-            pass
-        else:
-            print('Choix incorrect.')
-
-        for dossier in dossiers:
-            try:
-                shutil.rmtree(dossier)
-                print(f"Le dossier {dossier} a été supprimé.")
-            except OSError as e:
-                print("Error: %s - %s." % (e.foldername, e.strerror))
+        message = f'Voulez-vous supprimer tous les dossiers commençant par {self.livre.PREFIXE_DOSSIER_TEMPORAIRE} ? (O/N) '
+        dossiers = glob.glob(f'{self.livre.PREFIXE_DOSSIER_TEMPORAIRE}*')
+        if dossiers:
+            choix = demander_choix_binaire(message)
+            if choix:
+                brules = []
+                for dossier in dossiers:
+                    try:
+                        shutil.rmtree(dossier)
+                        brules.append(dossier)
+                    except OSError as e:
+                        print("Error: %s - %s." % (e.foldername, e.strerror))
+                print(f"Dossiers supprimés : {', '.join(brules)}.")
+            else:
+                print("Aucun dossier à supprimer.")
 
     def processed_data_to_csv(self, processed_data: dict[str, pd.DataFrame]):
         '''Enregistre les données traitées dans un fichier CSV.'''
